breaking down index into multiple files + utils
